Encoding POST bodies ourselves, instead of letting rest-client
It's a bit unfortunate we have to do this, but this avoids an even
more unfortunate interaction between rest-client and hashery that
prevents passing a hash as the payload.

Also update the tests appropriately. This is also a bit messy, but I
didn't want to pull in Rack just for a handful of tests.

Fixes #38