Manually override Ruby's SSL defaults when making Stripe requests
The Ruby maintainers are in the middle of a bit of a kerfluffle about
the SSL defaults (which are...fine as long as you're running a recent
Linux release or Mavericks and...aren't being man-in-the-middle'd)

As long as Ruby upstream is failing to take action, we're really in a
position where we should be protecting our users ourselves.

This overrides Ruby's default parameters for new SSL_CTX objects for
the duration of a Stripe API request. Specifically, it constrains the
cipher suite list and disables SSLv2, on the off chance it's
available.

Thanks to @jmhodges for making us aware of these specific issues, and
more generally for his noble crusade for better SSL configurations.