Add a length method to ListObject.
ListObjects behave much like arrays, and so I'd expect this method to exist. count is available via the api results, but it seems that always returns the full number of records, not the number of records in the current list. Yes, I could just use data.length instead, but better to take the path of least surprise.

I'd also recommend changing count to be the number of records returned, and change the existing attribute to something like total_count, as (at least in Ruby) count and length are synonyms.