Parameter encoding bug fix and clean-up
I believe that in #299 I introduced a bug that doesn't allow values
containing an ampersand to be properly encoded as described in #318.
(This is a little weird though, because as far as I can tell, the old
`URI.escape` didn't do the right thing there either.) This patch
implements a fix for that bug by dropping a custom escape implementation
and just using `URI.encode_www_form` to get the behavior we want.

Also introduces a number of tests to actually verify that the behavior
works as we expect.

One notable shift in implementation is that spaces in query parameters
are now encoded as `+` instead of `%20`. According to [1] this area is
still fairly poorly defined, but the `+` is more correct for query
values or form bodies (`%20` is more correct for URI paths, which is why
`URI.escape` was previously generating it for us). We should be fine
server side as I believe most HTTP implementations will tolerate either
at this point. See Rack's tests for that here for example [2].

[1] https://stackoverflow.com/questions/1634271/url-encoding-the-space-character-or-20
[2] https://github.com/rack/rack/blob/d938cb586cf2ac09521fc91e19edf77ecca99c1a/test/spec_utils.rb#L80,L81