Make StripeObjects immutable under some conditions
This has the effect of freezing StripeObjects under certain conditions
where they shouldn't be modified. The best (and currently only) example
of this is when an object is nested in an array of another resource, but
where it itself is not actually directly addressable in the API (meaning
that it can only be updated through its parent). Addressing objects for
update via array index is not stable so we take this immutable approach
where the whole array is replaced for a less volatile result. We can see
this at work in `account.legal_entities.additional_owners`.