Encode consistently between "create" and "update"
Looking into a reported `legal_entity` bug, I discovered that the
handling of a `nil` found in parameters for a create, e.g.

    account = Stripe::Account.create(country: nil)

Is fundamentally different than for an update:

    account.country = nil
    account.save

Essentially the internal behavior has always been to prune `nil`
values for our `.create` and `.list` methods, while on update, values
get run through an additional filter than converts `nil` to an empty
string (and are therefore not prune). It seems like this has been in
place for a while.

Sufficed to say, it would be nice to correct this incredibly
inconsistent behavior because it's illogical and can be very confusing
for anyone who accidentally comes across it. It seems like we should be
able to change `.create` and `.list` to behave the same as an update
here. There is a test in place to verify the pruning of `nil` values,
but it's not clear what would break if we stopped doing it (if
anything).

Another nice side effect of the change is that the `nil` to empty string
conversion gets moved to the encoding step, which would potentially
allow us to just send a `nil` for request encodes that support one (i.e.
`application/json`).