Explicitly set TLS 1.2 for OpenSSL versions that support it
This mirrors what we did for stripe/stripe-php#277 in that we explicitly
set TLS 1.2 to use for connections when we know that the Ruby's OpenSSL
supports it (i.e. greater or equal to OpenSSL 1.0.1a).

I'm not going to bring this in quite yet because so far we only have a
single user experiencing a problem, and it doesn't seem like they should
necessarily be seeing it given that they're on quite a recent operating
system.

Fixes #447.