Disallow TLS 1.0 using a cipher list
Disallows TLS 1.0 and some older protocols through the use of an OpenSSL
cipher list. This is an alternative to #448 and I believe it to be
better because it allows for forward compatibility in that we keep our
`ssl_version` flexibile (for TLS 1.3 once we get there).

See the comments in the patch for details, but we re-use whatever cipher
list would have already been in use today, but pack in a few additional
rules that ban old secure protocols.

There's some details on cipher lists here:

https://wiki.openssl.org/index.php/Manual:Ciphers(1)#CIPHER_STRINGS

rest-client has some useful comments on cipher lists here:

https://github.com/rest-client/rest-client/blob/3a79728b332d316afa281d10022658735a6f72df/lib/restclient/request.rb#L55,L114

And Ruby's default cipher list is here (but of course this doesn't apply
to older versions of Ruby):

https://github.com/ruby/ruby/blob/ea31b6c0e47ebb40769abe835a15d5e436f841be/ext/openssl/lib/openssl/ssl.rb#L22,L56