Make deep_copy public
It's helpful for our users to have this method public. Having a canonical way to duplicate/copy
an object so you can mutate it without fear will be helpful. DM me internally for example code.

Quick use case: pricing tiers contain `inf` when *submitting* the tiers to the API, but use `null` in the return value.
This makes comparing them challenging. To compare, you need to translate null => inf (or the other way around) and then do
a hash comparison. Creating a deep copy of the tiers/price object makes this operation more safe.